+++ weblate/trans/machine/base.py
@@ -74,7 +74,7 @@ class MachineTranslation(object):
         '''
         return
 
-    def json_req(self, url, http_post=False, skip_auth=False, **kwargs):
+    def json_req(self, url, http_post=False, skip_auth=False, raw=False, **kwargs):
         '''
         Performs JSON request.
         '''
@@ -124,6 +124,9 @@ class MachineTranslation(object):
         # Needed for Google
         while ',,' in text or '[,' in text:
             text = text.replace(',,', ',null,').replace('[,', '[')
+
+        if raw:
+            return text
 
         # Parse JSON
         response = json.loads(text)