+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2016.
 * Added quality check to check whether plurals are translated differently.
 * Fixed GitHub hooks for repositories with authentication.
 * Added optional Git exporter module.
+* Support for Microsoft Cognitive Services Translator API.
 
 weblate 2.9
 -----------